adding ability to register
